requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
anthropic
instructor
python-dotenv
slowapi
